Extract tokenization logic into token_matcher.rs
